requirements.txt
paho-mqtt
influxdb-client
PyYAML
streamlit
pandas
python-dotenv